+++ sample_app/lib/app.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: avoid_redundant_argument_values
+
 import 'dart:async';
 
 import 'package:firebase_core/firebase_core.dart';